Bugs fixing

+++ app/admin/artists.rb
@@ -37,10 +37,17 @@ permit_params :name, :description, :style, :image, :video, :spotify,
     link_to "retour", admin_root_path
   end
 
-   member_action :delete_artist_press_file, method: :delete do
+  member_action :delete_artist_press_file, method: :delete do
     artist = Artist.find(params[:id])
     file = artist.press_file
     file.purge_later
+    redirect_to edit_admin_artist_path
+  end
+
+  member_action :delete_artist_artist_video, method: :delete do
+    artist = Artist.find(params[:id])
+    vid = artist.artist_video
+    vid.purge_later
     redirect_to edit_admin_artist_path
   end
 
@@ -73,7 +80,10 @@ permit_params :name, :description, :style, :image, :video, :spotify,
       f.input :show_image3, as: :file, label: "Image 3"
       f.input :show_image4, as: :file, label: "Image 4"
       f.input :video
-      f;input :artist_video, as: :file
+      f.input :artist_video, as: :file
+      if artist.artist_video.attached?
+        span link_to "Supprimer",delete_artist_artist_video_admin_artist_path(artist),method: :delete,data: { confirm: 'Are you sure?' }
+      end
       f.input :spotify
       f.input :youtube
       f.input :insta
@@ -101,7 +111,9 @@ permit_params :name, :description, :style, :image, :video, :spotify,
       end
     end
       f.input :press_file, as: :file, label: "Dossier de presse"
+      if artist.press_file.attached?
         span link_to "Supprimer",delete_artist_press_file_admin_artist_path(artist),method: :delete,data: { confirm: 'Are you sure?' }
+      end
       f.input :active
       f.actions
     end
